pyproject.toml
[project]
name = "nova-code"
version = "0.1.0"
description = ""
authors = [
    {name = "Petro Kvartsianyi",email = "kvartsaniy@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi[uvicorn] (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "geoalchemy2 (>=0.17.0,<0.18.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "pydantic-extra-types (>=2.10.2,<3.0.0)",
    "overpy (>=0.7,<0.8)",
    "shapely (>=2.0.6,<3.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "greenlet (>=3.1.1,<4.0.0)",
    "aiohttp (>=3.11.11,<4.0.0)",
    "jinja2 (>=3.1.5,<4.0.0)",
    "aiocache (>=0.12.3,<0.13.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
